// src/api/models/ReleaseNote.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReleaseNote = {
  releaseId?: string | null;
  applicationName: string;
  version?: string | null;
  title: string;
  body: string;
  tags?: Array<string> | null;
  createdDate: string;
};